Yak shaving to squeeze a bit more speed
Mostly uses array args variants of several subs
to reduce subroutine overheads.

Identified using code profiling on a larger data set than previously used.

+++ lib/Biodiverse/Randomise.pm
@@ -8,6 +8,15 @@ use 5.010;
 
 use English qw / -no_match_vars /;
 
+#  a little debug
+#my $binsearch_gives_undef = 0;
+#END {
+#    say " ===== ";
+#    say "BINSEARCH WAS UNDEF:  $binsearch_gives_undef";
+#    say " ===== ";
+#}
+
+
 #use Devel::Symdump;
 use Data::Dumper qw { Dumper };
 use Carp;
@@ -734,7 +743,7 @@ sub get_randomised_basedata {
     @const_label_hash{@$constant_labels} = undef;
     for my $label ($bd->get_labels) {
         no autovivification;
-        my $groups = $bd->get_groups_with_label_as_hash (label => $label);
+        my $groups = $bd->get_groups_with_label_as_hash_aa ($label);
 
         #  we should cache the constant BD
         my $target_bd = exists $const_label_hash{$label} ? $const_bd : $non_const_bd;
@@ -1671,7 +1680,7 @@ END_PROGRESS_TEXT
         ###  - no point aiming for those that have it already
         ###  call will croak if label does not exist, so default to a blank hash
         my $new_bd_has_label
-            = eval {$new_bd->get_groups_with_label_as_hash (label => $label)}
+            = eval {$new_bd->get_groups_with_label_as_hash_aa ($label)}
             || {};
 
         #  cannot use $cloned_bd here, as it may not have the full set of groups yet
@@ -1698,7 +1707,7 @@ END_PROGRESS_TEXT
         ###  get the remaining original groups containing the original label.
         ###  Make sure it's a copy
         my %tmp
-            = $cloned_bd->get_groups_with_label_as_hash (label => $label);
+            = $cloned_bd->get_groups_with_label_as_hash_aa ($label);
         my $tmp_rand_order = $rand->shuffle ([sort keys %tmp]);
 
         my (
@@ -1789,8 +1798,12 @@ END_PROGRESS_TEXT
                 my $from_group = shift @$tmp_rand_order;
                 my $count = $tmp{$from_group};
 
-                delete $target_groups_hash{$to_group};
-                $self->delete_from_sorted_list_aa ($to_group, \@target_groups);
+                #  profiling suggests we get many $to_groups that are not in these lists,
+                #  so avoid some sub calls to save time 
+                if (exists $target_groups_hash{$to_group}) {
+                    $self->delete_from_sorted_list_aa ($to_group, \@target_groups);
+                    delete $target_groups_hash{$to_group};
+                }
 
                 warn "SELECTING GROUP THAT IS ALREADY FULL $to_group,"
                      . "$filled_groups{$to_group}, $target_richness{$to_group}, "
@@ -2367,12 +2380,12 @@ sub swap_to_reach_richness_targets {
         my $add_label = $labels[$i];
         
         
-        my $from_groups_hash = $cloned_bd->get_groups_with_label_as_hash (
-            label => $add_label,
-        );
+        my $from_groups_hash
+          = $cloned_bd->get_groups_with_label_as_hash_aa ($add_label);
+
         my $from_cloned_groups_tmp_a = $cloned_bd_groups_with_label_a{$add_label};
         if (!$from_cloned_groups_tmp_a  || !scalar @$from_cloned_groups_tmp_a) {
-            my $gps_tmp = $cloned_bd->get_groups_with_label_as_hash (label => $add_label);
+            my $gps_tmp = $cloned_bd->get_groups_with_label_as_hash_aa ($add_label);
             $from_cloned_groups_tmp_a = $cloned_bd_groups_with_label_a{$add_label} = [sort keys %$gps_tmp];
         };
 
@@ -2483,18 +2496,14 @@ sub swap_to_reach_richness_targets {
                 #  (Just use the first one).
                 my $old_gps_with_remove_label = $orig_bd_groups_with_label_a{$remove_label};
                 if (!$old_gps_with_remove_label) {  #  These do not change so access and cache.  Sort is for repeatability.
-                    my $gps = $bd->get_groups_with_label_as_hash (
-                        label => $remove_label,
-                    );
+                    my $gps = $bd->get_groups_with_label_as_hash_aa ($remove_label);
                     my @gps = sort keys %$gps;
                     $old_gps_with_remove_label = \@gps;
                     $orig_bd_groups_with_label_a{$remove_label} = $old_gps_with_remove_label;
                 }
 
                 my $cloned_self_gps_with_label
-                    = $cloned_bd->get_groups_with_label_as_hash (
-                        label => $remove_label,
-                    );
+                    = $cloned_bd->get_groups_with_label_as_hash_aa ($remove_label);
 
                 #  make sure it does not add to an existing case
                 #delete @old_groups{keys %$cloned_self_gps_with_label};
@@ -2534,17 +2543,13 @@ sub swap_to_reach_richness_targets {
                 $i = int $rand->rand (scalar @$unfilled_aref);
                 my $return_gp = $unfilled_aref->[$i];
 
-                $new_bd->add_element   (
-                    label => $remove_label,
-                    group => $return_gp,
-                    count => $removed_count,
-                    csv_object => $csv_object,
+                $new_bd->add_element_simple_aa (
+                    $remove_label,  $return_gp,
+                    $removed_count, $csv_object,
                 );
                 $swap_insert_count++;
 
-                my $new_richness = $new_bd->get_richness (
-                    element => $return_gp,
-                );
+                my $new_richness = $new_bd->get_richness_aa ($return_gp);
 
                 warn "ISSUES WITH RETURN $return_gp\n"
                   if $new_richness > $target_richness{$return_gp};
@@ -2589,17 +2594,15 @@ sub swap_to_reach_richness_targets {
 
             $swap_out_count ++;
 
-            if (!($swap_out_count % 5000)) {
+            if (!($swap_out_count % 10000)) {
                 say "Swap count $swap_out_count";
             }
         }
 
         #  add the new label to new_bd
-        $new_bd->add_element (
-            label => $add_label,
-            group => $target_group,
-            count => $add_count,
-            csv_object => $csv_object,
+        $new_bd->add_element_simple_aa (
+            $add_label, $target_group,
+            $add_count, $csv_object,
         );
         $swap_insert_count++;
         if (my $aref = $groups_without_labels_a{$add_label}) {
@@ -2615,7 +2618,7 @@ sub swap_to_reach_richness_targets {
         }
 
         #  check if we've filled this group, if nothing was swapped out
-        my $new_richness = $new_bd->get_richness (element => $target_group) // 0;
+        my $new_richness = $new_bd->get_richness_aa ($target_group) // 0;
 
         warn "ISSUES WITH TARGET $target_group\n"
           if $new_richness > $target_richness{$target_group};
@@ -2911,6 +2914,9 @@ sub delete_from_sorted_list {
     if (defined $idx) {
         splice @$list, $idx, 1;
     }
+    #else {
+    #    $binsearch_gives_undef++;
+    #}
 
     # skip the explicit return as a minor speedup for pre-5.20 systems
     $idx;
